+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Prefer native executable on Windows instead of lspjs
+
 ## v1.9.0 - 2024-08-28
 
 ### Added